Change float to decimal in json, remove references to custom primitives

+++ include/hirzel/json/JsonTokenType.hpp
@@ -12,7 +12,7 @@ namespace hirzel::json
 		Comma,
 		Colon,
 		String,
-		Float,
+		Decimal,
 		Integer,
 		True,
 		False,
